pyproject.toml
[project]
name = "multi-gpu-raster"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "hydra-core>=1.3.2",
    "lightning>=2.5.1.post0",
    "rasterio>=1.4.3",
    "shapely>=2.1.0",
    "torch>=2.7.0",
    "torchmetrics>=1.7.1",
    "torchvision>=0.22.0",
]

[dependency-groups]
dev = [
    "black>=25.1.0",
    "isort>=6.0.1",
    "pytest>=8.3.5",
]

[tool.pytest.ini_options]
addopts = [
    "-x",
    "--color=yes",
    "--verbose",
    "--durations=0"
]

[project.scripts]
mgpr = "multi_gpu_raster:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

 [tool.hatch.build.targets.wheel]
 packages = ["src/multi_gpu_raster"]